requirements.txt
aiodns==2.0.0
aiohttp==3.0.9
aiohttp-requests==0.1.2
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==19.1.0
beautifulsoup4==4.6.0
cchardet==2.1.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
cx-Oracle==7.1.3
fonttools==3.41.0
idna==2.8
idna-ssl==1.1.0
lxml==4.3.3
multidict==4.5.2
numpy==1.16.4
olefile==0.46
opencv-python==3.4.0.12
Pillow==4.1.1
pycares==3.0.0
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.6.1
redis==2.10.6
redis-py-cluster==1.3.6
requests==2.21.0
six==1.12.0
typing==3.6.6
urllib3==1.24.3
yarl==1.3.0
